# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: extraGO — Workforce Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: freelancerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: applyToJob
      tags: [applications]
      summary: Apply to a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/approve:
    post:
      operationId: approveApplication
      tags: [applications]
      summary: Approve an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/reject:
    post:
      operationId: rejectApplication
      tags: [applications]
      summary: Reject an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # WALLET
...
  /applications/{id}/complete:
    post:
      operationId: completeApplication
      tags: [applications]
      summary: Mark an application as completed and trigger payment cascade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/counter-offer:
    patch:
      operationId: counterOfferApplication
      tags: [applications]
      summary: Gold/Elite freelancer proposes a counter-offer rate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterOfferInput"
      responses:
        "200":
          description: Counter-offer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/accept-counter:
    post:
      operationId: acceptCounterOffer
      tags: [applications]
      summary: Company accepts a freelancer counter-offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Counter-offer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/reject-counter:
    post:
      operationId: rejectCounterOffer
      tags: [applications]
      summary: Company rejects a freelancer counter-offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Counter-offer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [company, freelancer, admin]
        adminRole:
          type: ["string", "null"]
          description: Sub-role for admin users (super_admin, finance_admin, etc.)
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        pixKey:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        level:
          type: string
          enum: [bronze, silver, gold, elite, diamond]
        reputationScore:
          type: number
        completedJobs:
          type: integer
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        profileCompletion:
          type: integer
        referralCode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - level
        - reputationScore
        - completedJobs
        - isVerified
        - isBanned
        - profileCompletion
        - referralCode
        - createdAt
        - categories
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        workersNeeded:
          type: integer
        workersApproved:
          type: integer
        hourlyRate:
          type: number
        totalValue:
          type: number
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        companyId:
          type: integer
        companyName:
          type: ["string", "null"]
        companyAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - location
        - date
        - startTime
        - endTime
        - workersNeeded
        - workersApproved
        - hourlyRate
        - totalValue
        - status
        - companyId
        - createdAt
...
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        workersNeeded:
          type: integer
        hourlyRate:
          type: number
    Application:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        freelancerId:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, completed, cancelled, counter_offered, counter_accepted, counter_rejected]
        message:
          type: ["string", "null"]
        proposedRate:
          type: ["number", "null"]
        appliedAt:
          type: string
        job:
          $ref: "#/components/schemas/Job"
        freelancer:
          $ref: "#/components/schemas/User"
      required:
        - id
        - jobId
        - freelancerId
        - status
        - appliedAt
    CounterOfferInput:
      type: object
      required: [proposedRate]
      properties:
        proposedRate:
          type: number
    ApplicationInput:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer
        message:
          type: ["string", "null"]
...